genai/requirements.txt
Flask>=2.0.0
python-dotenv>=1.0.0
requests>=2.25.0
google-generativeai>=0.3.0